// filmosearch.frontend/src/api/api.ts
export interface ActorsLookupDto {
    id?: number;
    firstName?: string | undefined;
    lastName?: string | undefined;
}

export interface AddFilmToActorDto {
    actorFirstName?: string | undefined;
    actorLastName?: string | undefined;
    filmTitle?: string | undefined;
}

export interface CreateActorDto {
    firstName?: string | undefined;
    lastName?: string | undefined;
}

export interface CreateFilmDto {
    title?: string | undefined;
}
// ...
export interface FilmDetailsVm {
    id?: number;
    title?: string | undefined;
    actors?: ActorsLookupDto[] | undefined;
    reviews?: ReviewsLookupDto[] | undefined;
}

export interface FilmListVm {
    films?: FilmLookupDto[] | undefined;
}

export interface FilmLookupDto {
    id?: number;
    title?: string | undefined;
}
// ...
export interface ReviewsLookupDto {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    stars?: number;
}

export interface UpdateActorDto {
    id?: number;
    firstName?: string | undefined;
    lastName?: string | undefined;
}

export interface UpdateFilmDto {
    id?: number;
    title?: string | undefined;
}